Fix link-time errors on Linux; improve Makefile
The original build recipe gave lots of undefined symbols at link time
(building on Linx Mint 21.1 Vera); it turned out that the linker (GNU
ln) loads object files and libraries in the order they're given on the
command line.  That means that the libraries have to come _after_ the
source files on the gcc command line.

The implicit rule for building from C sources uses two make variables,
CFLAGS at the beginning of the line, and LDLIBS at the end.  So I
added those to the Makefile, along with the conventional CC for the C
compiler.  These changes will almost certainly work on Windows and Mac;
I don't think I used any GNU make features.  But I haven't tested it.

This commit also adds a local `.gitignore` file to keep the binary file
from being accidentally added to the repository.